# IEUM Chain v0.0.6.3 RPC 컴파일 수정 RPC 모듈은 `u128` 금액과 체크포인트·블록 조회 기능을 요구하지만 기존 원장 핵심 자료형은 여전히 `u64`이고 필요한 `Blockchain` API가 없어 컴파일 오류가 발생했습니다.

+++ src/model.rs
@@ -8,8 +8,9 @@ pub type Address = String;
 pub struct Transaction {
     pub from: Address,
     pub to: Address,
-    pub amount: u64,
-    pub fee: u64,
+    /// 1 IEUM = 10^18 최소 단위이므로 전체 발행량을 안전하게 담도록 u128을 사용합니다.
+    pub amount: u128,
+    pub fee: u128,
     pub nonce: u64,
     pub signature: String,
 }
